requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
billiard==3.6.4.0
celery==4.4.7
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Deprecated==1.2.13
Django==2.2
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-ckeditor==6.3.2
django-cors-headers==3.11.0
django-js-asset==2.0.0
django-timezone-field==4.2.3
dnspython==2.2.1
eventlet==0.33.1
greenlet==1.1.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==4.6.11
msgpack==1.0.3
packaging==21.3
Pillow==9.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
pypiwin32==223
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
redis==4.3.1
requests==2.27.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing-extensions==4.2.0
urllib3==1.26.9
vine==1.3.0
wrapt==1.14.1
zope.interface==5.4.0
